Add necessary scripts for travis-ci and update gems
The original .travis.yml file was configured for Node, which was a
mistake. The new file is now setup for Jekyll, which uses Ruby. All it
does is test the Jekyll build command. Gems were also updated, and a
script was added to the /bin directory so that travis-ci can test the
jekyll build properly.